requirements.txt
aiogram==3.15.0
aiohttp==3.9.5
python-dotenv==1.0.1
redis==5.2.1
pillow==11.0.0